+++ rstar/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Unreleased
+## Changed
+- Removed dependency on `pdqselect` ([PR](https://github.com/georust/rstar/pull/85))
+- New **minimal supported rust version (MSRV): 1.49.0**
 
 # 0.9.2
 - Add `RTree::drain_*` methods to remove and drain selected items. ([PR](https://github.com/georust/rstar/pull/77))