+++ rstar/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased
 ## Changed
 - Removed dependency on `pdqselect` ([PR](https://github.com/georust/rstar/pull/85))
-- New **minimal supported rust version (MSRV): 1.51.0**
+- BREAKING: New **minimal supported rust version (MSRV): 1.51.0**
 - Replace all usages of `std` with `core` & `alloc` to make `rstar` fit for
   `no_std`. ([PR](https://github.com/georust/rstar/pull/83))
 - Updated `heapless` dependency to 0.7 to make use of const generics. ([PR](https://github.com/georust/rstar/pull/87))